updated button placement
